refactor(worker): isolate workspace runner
